+++ eitprocessing/datacollection/loading/binreader.py
