+++ eitprocessing/datacollection/loading/load_timpel.py
@@ -6,13 +6,13 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 
-from eitprocessing.continuous_data import ContinuousData
-from eitprocessing.data_collection import DataCollection
-from eitprocessing.eit_data import EITData
-from eitprocessing.eit_data.loading import load_data
-from eitprocessing.eit_data.phases import MaxValue, MinValue, QRSMark
-from eitprocessing.eit_data.vendor import Vendor
-from eitprocessing.sparse_data import SparseData
+from eitprocessing.datacollection import DataCollection
+from eitprocessing.datacollection.continuousdata import ContinuousData
+from eitprocessing.datacollection.eitdata import EITData
+from eitprocessing.datacollection.loading import load_data
+from eitprocessing.datacollection.phases import MaxValue, MinValue, QRSMark
+from eitprocessing.datacollection.sparsedata import SparseData
+from eitprocessing.datacollection.vendor import Vendor
 
 if TYPE_CHECKING:
     from pathlib import Path
