+++ eitprocessing/datacollection/sparsedata.py
